feat(design-parity): render the Device screen on the CMP desktop backend
Move the presentational DeviceBody (+ LastMessageInfo, the section helpers, and
the two parity @Previews) out of :app into :meshcore-components commonMain, so
the candidate renders on the CMP desktop (Skiko) backend instead of Android/
Robolectric. The stateful DeviceScreen/ConnectedDevice wrappers stay in :app.

- DeviceBody.kt + DeviceBodyPreviews.kt in commonMain; icons swapped to the
  vendored MeshIcons; COMMANDS_CHANNEL_NAME moved alongside; app files repointed.
- Apply id("ee.schimke.composeai.preview") to :meshcore-components — the CLI
  can't auto-inject into a com.android.kotlin.multiplatform.library module.
- design-map.json + the HTML exports' componentId point at the components
  handles; the workflow packs --module meshcore-components.

End-to-end: bundle pack reports backend=desktop, and design-parity pairs both
themes (light ~11%, dark ~6% over the overlap) — comparable to the Android path,
confirming the desktop render (bundled fonts + vendored icons + MeshCore theme)
is faithful. Verified :app:compileDebugKotlin + components desktop/android green.
