+++ app/src/main/kotlin/ee/schimke/meshcore/app/ui/CachedDeviceScreen.kt
@@ -1,5 +1,6 @@
 package ee.schimke.meshcore.app.ui
 
+import ee.schimke.meshcore.components.ui.DeviceBody
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
