+++ design/DeviceScreen.dark.html
@@ -274,7 +274,7 @@
 
     <script type="application/design-parity+json">
       {
-        "componentId": "app/src/main/kotlin/ee/schimke/meshcore/app/ui/DeviceScreenPreviews.kt#DeviceBodyDarkPreview",
+        "componentId": "meshcore-components/src/commonMain/kotlin/ee/schimke/meshcore/components/ui/DeviceBodyPreviews.kt#DeviceBodyDarkPreview",
         "tokens": {
           "spacing": { "screenPadding": 16, "cardGap": 12, "rowGap": 8, "cardPadding": 16 },
           "radius": { "card": 12, "chip": 8, "avatar": 18 },
