+++ gradle/libs.versions.toml
@@ -10,6 +10,7 @@ androidx-lifecycle = "2.10.0"
 androidx-testExt = "1.3.0"
 composeMultiplatform = "1.11.0"
 composePreviewAnnotations = "0.14.0"
+composePreviewPlugin = "0.15.1"
 core = "1.7.0"
 kotlin = "2.3.21"
 kotlinx-coroutines = "1.11.0"
@@ -137,3 +138,4 @@ aboutlibraries = { id = "com.mikepenz.aboutlibraries.plugin", version.ref = "abo
 ktfmt = { id = "com.ncorti.ktfmt.gradle", version.ref = "ktfmt" }
 playPublisher = { id = "com.github.triplet.play", version.ref = "playPublisher" }
 tapmoc = { id = "com.gradleup.tapmoc", version.ref = "tapmoc" }
+composePreview = { id = "ee.schimke.composeai.preview", version.ref = "composePreviewPlugin" }
